requirements.txt
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
attrs==24.2.0
blinker==1.9.0
build==1.2.2.post1
CacheControl==0.14.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cleo==2.1.0
click==8.1.7
comm==0.2.2
crashtest==0.4.1
cryptography==43.0.3
debugpy==1.8.8
decorator==5.1.1
distlib==0.3.9
dulwich==0.21.7
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.4
fastjsonschema==2.20.0
filelock==3.16.1
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.67.1
grpcio-tools==1.67.1
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
installer==0.7.0
ipykernel==6.29.5
ipython==8.29.0
jaraco.classes==3.4.0
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keyring==24.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.5.0
msgpack==1.1.0
narwhals==1.13.3
nest-asyncio==1.6.0
numpy==2.1.3
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
pkginfo==1.11.2
platformdirs==4.3.6
poetry==1.8.4
poetry-core==1.9.1
poetry-plugin-export==1.8.0
prompt_toolkit==3.0.48
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-multipart==0.0.17
pytz==2024.2
pyzmq==26.2.0
RapidFuzz==3.10.1
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.21.0
scikit-learn==1.5.2
scipy==1.14.1
SecretStorage==3.3.3
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
stack-data==0.6.2
starlette==0.41.2
streamlit==1.40.1
tenacity==9.0.0
threadpoolctl==3.5.0
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.1
traitlets==5.14.3
trove-classifiers==2024.10.21.16
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
virtualenv==20.27.1
watchdog==6.0.0
wcwidth==0.2.13
wheel==0.44.0
zipp==3.20.2
